+++ src/components/PageNotFound.svelte
@@ -6,7 +6,7 @@
   import { Route } from 'svelte-navigator'
 
   // Scripts
-  import { errorSeparator } from '../Products/services/productsService'
+  import { errorSeparator } from '../products/services/productsService'
   import { title } from '../stores/title'
 
   const message: string = `001${errorSeparator}¡Estamos perdidos!${errorSeparator}Parece que nos has solicitado una página que no existe. Haz click en el botón  para darte una vuelta por nuestro catálogo y hacer tu primera compra con nosotros.`
