+++ src/products/components/CartProductItem.svelte
