+++ src/products/components/ProductItem.svelte
