+++ src/products/components/ProductsLoader.svelte
