+++ src/products/components/SingleProduct.svelte
